refactor: change composition root and orchestrator structure, dedicate bump use case, cleanup code

+++ src/Calaf/Application/Ports/Git.fs
@@ -4,7 +4,7 @@ open Calaf.Contracts
 
 type IGit =
     abstract tryRead:
-        directory: string -> maxTagsToRead: int -> tagsPrefixesToFilter: string list -> timeStamp: System.DateTimeOffset -> Result<GitRepositoryInfo option, CalafError>
+        directory: string -> maxTagsToRead: byte -> tagsPrefixesToFilter: string list -> timeStamp: System.DateTimeOffset -> Result<GitRepositoryInfo option, CalafError>
         
     abstract tryApply:
         directory: string -> commitMessage: string -> tagName: string -> signature: GitSignatureInfo -> Result<unit, CalafError>
